Move error functions from util.c to error.c
