+++ Makefile
@@ -3,7 +3,7 @@ CC=gcc
 CFLAGS=-std=gnu99 -Wall -Wextra -ggdb3
 LDLIBS=-lm
 
-SRCS=main.c expr.c env.c lambda.c util.c debug.c read.c lexer.c parser.c eval.c prim_special.c prim_general.c prim_logic.c prim_type.c prim_list.c prim_string.c prim_arith.c prim_bitwise.c prim_io.c
+SRCS=main.c expr.c env.c lambda.c util.c error.c debug.c read.c lexer.c parser.c eval.c prim_special.c prim_general.c prim_logic.c prim_type.c prim_list.c prim_string.c prim_arith.c prim_bitwise.c prim_io.c
 OBJS=$(addprefix obj/, $(addsuffix .o, $(SRCS)))
 
 BIN=sl
